Adding redirect strategy so page can reload
