+++ apps/dashboard/src/app/login/login.component.ts
@@ -15,19 +15,7 @@ export class LoginComponent implements OnInit {
 
   constructor(private iam: Iam, private router: Router) { }
 
-  ngOnInit() {
-    const token = localStorage.getItem("token");
-    if (token !== null) {
-    this.iam.runExecutable("admin", "graph", "validate-token", [{token: token}])
-    .subscribe((result: any) => {
-      if (result.result.length > 0) {
-        const user = result.result[0];
-        this.iam.setUser(user.username, token);
-        this.router.navigate(["/home"]);
-      }
-    });
-    }
-  }
+  ngOnInit() { }
 
   login() {
     this.iam.runExecutable("admin", "graph", "gen-token", [{email: this.email}, {password: this.password}])
