+++ images/dashboard/Dockerfile
@@ -1,2 +1,2 @@
-FROM icanplayguitar94/iam:base-04ab437d8aaa5d83187c80d7fc114073a65be3bb
+FROM icanplayguitar94/iam:base-redirect
 COPY ./src public/dist
