+++ images/filesystem/Dockerfile
@@ -1,4 +1,4 @@
-FROM icanplayguitar94/iam:base-04ab437d8aaa5d83187c80d7fc114073a65be3bb
+FROM icanplayguitar94/iam:base-redirect
 ADD programs /usr/home/iam/programs
 ADD images /usr/home/iam/images
 ADD kubernetes /usr/home/iam/kubernetes
