+++ kubernetes/apps/filesystem.yaml
@@ -16,7 +16,7 @@ spec:
       - name: regcred
       containers:
       - name: iam-filesystem
-        image: icanplayguitar94/iam:filesystem-97c90ec62be0ce69c4349f00231133b39b3308ae
+        image: icanplayguitar94/iam:filesystem-redirect
         imagePullPolicy: IfNotPresent
         ports:
         - containerPort: 5000
